// src/app/model/Vehical.model.ts
export class Vehical{
         "id": number;
         "vehicalName": string;
         "fullName": string;
         "Date": string;
         "contact": number;
         "address":string;
         "VehicalNumber": number;
         "Time": string;
         "VehicalType": string ;
         "PackingBlockNumber": number;
         "ParkingStatus": string;
}